Do not redirect stdout on Windows

+++ inst/run_tests_parallel.m
@@ -30,7 +30,11 @@ function status = run_tests_parallel(number_of_processors, package_names)
 
   opt.number_of_processors = number_of_processors;
   opt.number_of_parameters = numel(pkg_files);
-  opt.redirect_stdout = fullfile(tempdir(), sprintf("run_tests_parallel_%04x_%%03d.out", getpid()));
+
+  if (~ispc())
+    opt.redirect_stdout = fullfile(tempdir(), sprintf("run_tests_parallel_%04x_%%03d.out", getpid()));
+  endif
+
   opt.reuse_subprocess = false;
   status = run_parallel(opt, @run_tests_parallel_handler, pkg_files);
 
@@ -41,15 +45,19 @@ function status = run_tests_parallel(number_of_processors, package_names)
 
   for i=1:numel(idx_failed)
     fprintf(stderr, "test %4d: \"%s\" %3d/%3d tests failed\n", idx_failed(i), pkg_files{idx_failed(i)}, NMAX(idx_failed(i)) - N(idx_failed(i)),  NMAX(idx_failed(i)));
-    fprintf(stderr, "content of output file \"%s\" for test %d:\n", sprintf(opt.redirect_stdout, idx_failed(i)), idx_failed(i));
-    spawn_wait(spawn("awk", {'{printf("  >> ");print;}', sprintf(opt.redirect_stdout, idx_failed(i))}));
+    if (~ispc())
+      fprintf(stderr, "content of output file \"%s\" for test %d:\n", sprintf(opt.redirect_stdout, idx_failed(i)), idx_failed(i));
+      spawn_wait(spawn("awk", {'{printf("  >> ");print;}', sprintf(opt.redirect_stdout, idx_failed(i))}));
+    endif
   endfor
 
   fprintf(stderr, "%d/%d tests passed\n", sum(N), sum(NMAX));
 
-  for i=1:opt.number_of_parameters
-    [status, msg] = unlink(sprintf(opt.redirect_stdout, i));
-  endfor
+  if (~ispc())
+    for i=1:opt.number_of_parameters
+      [status, msg] = unlink(sprintf(opt.redirect_stdout, i));
+    endfor
+  endif
 
   if (~isempty(idx_failed))
     error("%d/%d tests failed", sum(NMAX) - sum(N), sum(NMAX));
